Moved Files in Folder
I have directed the files into this src folder to make some breathing room. I have added a keyPress file where it records your key presses and submits it into the terminal. I made a file for capturing a video on the drone camera and main was just moved without any changes to it yet.