Updates and Successful test!!
'theWatcher' is working correctly, the test went very well as I flew the drone via keyboard therefore made an RC Drone. There are times where the value gets stuck into the same one. for example, when I press 'w' to go forward, it stays that and I have to make it go back or land if it were to get near any obsticle or person. I have updated the battery module to be more specific and changed the screen size for the imageCapture to a different aspect ratio. I have completed many goals so I can be confident that the files work